+++ Sources/SwiftUIKit/_Deprecated/DocumentPresenter.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+#if os(iOS)
+@available(*, deprecated, renamed: DocumentGroupPresenter)
+public typealias DocumentPresenter = DocumentGroupPresenter
+#endif